Add simple script, make client arch-sensitive
