+++ scripts/run_all.sh
@@ -26,9 +26,9 @@
 set -xu
 
 GEM5=./../build/ALL/gem5.opt
-GEM5_CONFIG=./gem5-configs/fs-multiarch.py
+GEM5_CONFIG=./gem5-configs/fs-simple.py
 
-ARCH="arm64"
+ARCH="amd64"
 CPU_TYPE="o3"
 
 BENCHMARKS=()
